refactor: 결과 화면에서 결과 출력 형식 메시지를 만드는 포매터 책임을 분리
